Write the rest of segment registers in GDT

+++ kernel/src/interrupts.rs
@@ -7,7 +7,7 @@ lazy_static! {
         let mut idt = InterruptDescriptorTable::new();
         idt.breakpoint.set_handler_fn(breakpoint_handler);
         unsafe {
-            idt.double_fault.set_handler_fn(double_fault_handler).set_stack_index(gdt::DOUBLE_FAULT_IST_INDEX);
+            idt.double_fault.set_handler_fn(double_fault_handler).set_stack_index(gdt::DOUBLE_FAULT_IST_INDEX as u16);
         }
         idt.page_fault.set_handler_fn(page_fault_handler);
         idt
